class years_combine
ini kelas years_combine
