+++ TaxCalculator/src/lib/years_combine.java
@@ -1,0 +1,24 @@
+package lib;
+//noval fauzan_1302190046
+
+public class years_combine {
+    private int yearJoined;
+	private int monthJoined;
+	private int dayJoined;
+	private int monthWorkingInYear;
+
+
+public int getAnnualIncomeTax() {
+		
+		//Menghitung berapa lama pegawai bekerja dalam setahun ini, jika pegawai sudah bekerja dari tahun sebelumnya maka otomatis dianggap 12 bulan.
+		LocalDate date = LocalDate.now();
+		
+		if (date.getYear() == yearJoined) {
+			monthWorkingInYear = date.getMonthValue() - monthJoined;
+		}else {
+			monthWorkingInYear = 12;
+		}
+		
+		return TaxFunction.calculateTax(monthlySalary, otherMonthlyIncome, monthWorkingInYear, annualDeductible, spouseIdNumber.equals(""), childIdNumbers.size());
+	}
+}
